requirements.txt
playwright==1.42.0
python-dotenv
requests